github_agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
PyGithub>=2.1.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
markdownify>=0.11.0
click>=8.1.0
rich>=13.0.0
